move substitution to own file
